% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROlogit.R
\name{loglikhresid}
\alias{loglikhresid}
\title{The negative log likelihood function for obtaining heuristic residuals}
\usage{
loglikhresid(par, y, x, par1)
}
\arguments{
\item{par}{vector of numeric. Contain the intercept, first entry, and
coefficient of log(sigma), second entry.}

\item{y}{vector of numeric. Centered outcomes within each stratum.}

\item{x}{matrix. The design matrix including intercept term.}

\item{par1}{vector of numeric. The estimated coefficients from RO-logit
model.}
}
\description{
Compute the negative log-likelihood for obtaining heuristic
  residuals.
}
