% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROlogit.R
\name{qqplot.EVT1}
\alias{qqplot.EVT1}
\title{Make Q-Q plot for residual diagnostics.}
\usage{
qqplot.EVT1(hresid, scale)
}
\arguments{
\item{hresid}{vector of numeric. The heuristic residuals.}

\item{scale}{numeric. The scale parameter.}
}
\description{
Make Q-Q plot for residual diagnostics.
}
\examples{
\dontrun{
hresid <- evd::rgumbel(n = 100, loc = 0, scale = 3)
qqplot.EVT1(hresid = hresid, scale = 3)
}
}
