% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROlogit.R
\name{rologit.coxph}
\alias{rologit.coxph}
\title{Fit RO-logit model using Cox-PH}
\usage{
rologit.coxph(design.mat, y, svar = NULL, method = "efron")
}
\arguments{
\item{design.mat}{The design matrix without intercept term.}

\item{y}{vector of numeric. The outcomes.}

\item{svar}{vector of numeric. The strata variable.}

\item{method}{string. Use Efron (\code{"efron"}) or Breslow
(\code{"breslow"}) method for handling ties in the outcome. The default is
\code{"efron"}. See \code{\link[survival]{coxph}} for details.}
}
\value{
Returns the model fitted using \code{coxph}.
}
\description{
Fit RO-logit model using Cox-PH
}
