% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROlogit.R
\name{summary.rologit}
\alias{summary.rologit}
\title{Summarise RO-Logit Model}
\usage{
\method{summary}{rologit}(object, ...)
}
\arguments{
\item{object}{Model object fitted using \code{\link{rologit}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Prints the estimated coefficients of an RO-Logit model.
}
